+++ test/test_component/pch.h
@@ -1,5 +1,6 @@
 #pragma once
 
 #define WINRT_LEAN_AND_MEAN
+#define WINRT_NO_MAKE_DETECTION // for test in Class.cpp
 #include <numeric>
 #include <winrt/Windows.Foundation.h>
